refactor: rename cli.mts -> cli-parser.mts to clarify it is not the entrypoint
bin/index.mjs (via bin/pr-shepherd) is the executable; cli-parser.mjs is the
argument-parsing library imported by the entrypoint. The old name caused agents
to mistake bin/cli.mjs for the runnable CLI.
